more refactoring, split the functions again to files
