+++ app.go
@@ -1,37 +1,21 @@
 package xlog
 
 import (
-	"bytes"
-	"context"
 	"crypto/rand"
 	"embed"
 	"encoding/json"
-	"errors"
-	"flag"
 	"fmt"
 	"html/template"
-	"io"
 	"io/fs"
 	"log/slog"
 	"net/http"
-	"net/http/httptest"
 	"os"
 	"path"
-	"reflect"
 	"regexp"
 	"runtime"
-	"slices"
-	"strings"
 	"time"
 
-	"iter"
-	"sort"
-
-	"github.com/emad-elsaid/xlog/markdown/ast"
-	emojiAst "github.com/emad-elsaid/xlog/markdown/emoji/ast"
 	"github.com/gorilla/csrf"
-	"gitlab.com/greyxor/slogor"
-	"golang.org/x/sync/errgroup"
 )
 
 // Property represent a piece of information about the current page such as last
@@ -229,59 +213,6 @@ func (app *App) initDefaultHelpers() {
 	}
 }
 
-// Start initializes and starts the application
-func (app *App) Start(ctx context.Context) {
-	runtime.GOMAXPROCS(runtime.NumCPU() * 2)
-	flag.Parse()
-
-	// Setup logger
-	level := slogor.SetLevel(slog.LevelDebug)
-	timeFmt := slogor.SetTimeFormat(time.TimeOnly)
-	handler := slogor.NewHandler(os.Stderr, level, timeFmt)
-	logger := slog.New(handler)
-	slog.SetDefault(logger)
-
-	// if a static site is going to be built then lets also turn on read only
-	// mode
-	if len(app.config.Build) > 0 {
-		app.config.Readonly = true
-	}
-
-	if !app.config.Readonly {
-		app.Listen(PageChanged, func(p Page) error { app.clearPagesCache(p); return nil })
-		app.Listen(PageDeleted, func(p Page) error { app.clearPagesCache(p); return nil })
-	}
-
-	if err := os.Chdir(app.config.Source); err != nil {
-		slog.Error("Failed to change dir to source", "error", err, "source", app.config.Source)
-		os.Exit(1)
-	}
-
-	app.initExtensions()
-
-	app.Get("/{$}", app.rootHandler)
-	app.Get("/{page...}", app.getPageHandler)
-
-	if len(app.config.Build) > 0 {
-		if err := app.build(app.config.Build); err != nil {
-			slog.Error("Failed to build static pages", "error", err)
-			os.Exit(1)
-		}
-
-		return
-	}
-
-	srv := app.server()
-	slog.Info("Starting server", "address", app.config.BindAddress)
-
-	go func() {
-		<-ctx.Done()
-		srv.Close()
-	}()
-
-	srv.ListenAndServe()
-}
-
 // server creates and configures the HTTP server
 func (app *App) server() *http.Server {
 	app.compileTemplates()
@@ -352,128 +283,14 @@ func (app *App) handlerFuncToHttpHandler(handler HandlerFunc) http.HandlerFunc {
 	}
 }
 
-// RegisterExtension registers a new extension
-func (app *App) RegisterExtension(e Extension) {
-	app.extensions = append(app.extensions, e)
-}
-
-// RegisterWidget registers a widget function
-func (app *App) RegisterWidget(s WidgetSpace, priority float32, f WidgetFunc) {
-
-	pl, ok := app.widgets[s]
-	if !ok {
-		pl = new(priorityList[WidgetFunc])
-		app.widgets[s] = pl
-	}
-
-	pl.Add(f, priority)
-}
-
-// RegisterCommand registers a new command
-func (app *App) RegisterCommand(c func(Page) []Command) {
-	app.commands = append(app.commands, c)
-}
-
-// RegisterQuickCommand registers a new quick command
-func (app *App) RegisterQuickCommand(c func(Page) []Command) {
-	app.quickCommands = append(app.quickCommands, c)
-}
-
-// RegisterLink registers a new link
-func (app *App) RegisterLink(l func(Page) []Command) {
-	app.links = append(app.links, l)
-}
-
 // RegisterPageSource registers a page source
 func (app *App) RegisterPageSource(p PageSource) {
 	app.sources = append([]PageSource{p}, app.sources...)
 }
 
-// RegisterPreprocessor registers a preprocessor function
-func (app *App) RegisterPreprocessor(f Preprocessor) {
-	app.preprocessors = append(app.preprocessors, f)
-}
-
-// RegisterProperty registers a function that returns a set of properties for a page
-func (app *App) RegisterProperty(a func(Page) []Property) {
-	app.propsSources = append(app.propsSources, a)
-}
-
 // RequireHTMX registers HTMX library
 func (app *App) RequireHTMX() {
 	app.includeJS("/public/htmx.min.js")
-}
-
-// PreProcess processes content through all registered preprocessors
-func (app *App) PreProcess(content Markdown) Markdown {
-
-	for _, v := range app.preprocessors {
-		content = v(content)
-	}
-
-	return content
-}
-
-// initExtensions initializes all registered extensions
-func (app *App) initExtensions() {
-	if app.config.DisabledExtensions == "all" {
-		slog.Info("extensions", "disabled", "all")
-		return
-	}
-
-	disabled := strings.Split(app.config.DisabledExtensions, ",")
-	disabledNames := []string{} // because the user can input wrong extension name
-	enabledNames := []string{}
-	for i := range app.extensions {
-		if slices.Contains(disabled, app.extensions[i].Name()) {
-			disabledNames = append(disabledNames, app.extensions[i].Name())
-			continue
-		}
-
-		app.extensions[i].Init()
-		enabledNames = append(enabledNames, app.extensions[i].Name())
-	}
-
-	slog.Info("extensions", "enabled", enabledNames, "disabled", disabled)
-}
-
-// RegisterTemplate registers a filesystem that contains templates
-func (app *App) RegisterTemplate(t fs.FS, subDir string) {
-
-	ts, _ := fs.Sub(t, subDir)
-	app.templatesFSs = append(app.templatesFSs, ts)
-}
-
-// RegisterStaticDir adds a filesystem to the static files list
-func (app *App) RegisterStaticDir(f fs.FS) {
-	app.staticDirs = append(app.staticDirs, f)
-}
-
-// RegisterHelper registers a new helper function
-func (app *App) RegisterHelper(name string, f any) error {
-
-	if _, ok := app.helpers[name]; ok {
-		return ErrHelperRegistered
-	}
-
-	app.helpers[name] = f
-	return nil
-}
-
-// IgnorePath registers a pattern to be ignored when walking directories
-func (app *App) IgnorePath(r *regexp.Regexp) {
-	app.ignoredPaths = append(app.ignoredPaths, r)
-}
-
-// IsIgnoredPath checks if a file path should be ignored
-func (app *App) IsIgnoredPath(d string) bool {
-
-	for _, v := range app.ignoredPaths {
-		if v.MatchString(d) {
-			return true
-		}
-	}
-	return false
 }
 
 // GetConfig returns the application configuration
@@ -486,65 +303,6 @@ func (app *App) GetRouter() *http.ServeMux {
 	return app.router
 }
 
-// compileTemplates compiles all registered templates
-func (app *App) compileTemplates() {
-	const ext = ".html"
-
-	// add default templates before everything else
-	sub, _ := fs.Sub(defaultTemplates, "templates")
-	app.templatesFSs = append([]fs.FS{sub}, app.templatesFSs...)
-	// add theme directory after everything else to allow user to override any template
-	if _, err := os.Stat("theme"); err == nil {
-		app.templatesFSs = append(app.templatesFSs, os.DirFS("theme"))
-	}
-
-	app.templates = template.New("")
-	for _, tfs := range app.templatesFSs {
-		fs.WalkDir(tfs, ".", func(p string, d fs.DirEntry, err error) error {
-			if err != nil {
-				return err
-			}
-
-			if strings.HasSuffix(p, ext) && d.Type().IsRegular() {
-				ext := path.Ext(p)
-				name := strings.TrimSuffix(p, ext)
-				slog.Info("Template " + name)
-
-				c, err := fs.ReadFile(tfs, p)
-				if err != nil {
-					return err
-				}
-
-				template.Must(app.templates.New(name).Funcs(app.helpers).Parse(string(c)))
-			}
-
-			return nil
-		})
-	}
-}
-
-// Partial executes a template by it's path name
-func (app *App) Partial(path string, data Locals) template.HTML {
-	v := app.templates.Lookup(path)
-	if v == nil {
-		return template.HTML(fmt.Sprintf("template %s not found", path))
-	}
-
-	if data == nil {
-		data = Locals{}
-	}
-
-	data["config"] = app.config
-
-	w := bytes.NewBufferString("")
-
-	if err := v.Execute(w, data); err != nil {
-		return template.HTML("rendering error " + path + " " + err.Error())
-	}
-
-	return template.HTML(w.String())
-}
-
 // requestLoggerHandler logs HTTP requests
 func (app *App) requestLoggerHandler(h http.Handler) http.Handler {
 	return http.HandlerFunc(func(w Response, r Request) {
@@ -554,254 +312,9 @@ func (app *App) requestLoggerHandler(h http.Handler) http.Handler {
 	})
 }
 
-// rootHandler redirects to the index page
-func (app *App) rootHandler(r Request) Output {
-	return app.Redirect("/" + app.config.Index)
-}
-
-// getPageHandler handles page requests
-func (app *App) getPageHandler(r Request) Output {
-	page := app.NewPage(r.PathValue("page"))
-
-	if page == nil {
-		return app.NoContent()
-	}
-
-	if !page.Exists() {
-		// if it's a directory get back to home page
-		if s, err := os.Stat(page.Name()); err == nil && s.IsDir() {
-			return app.Redirect(app.config.Index)
-		}
-
-		// if it's a static file serve it
-		if output, err := app.staticHandler(r); err == nil {
-			return output
-		}
-
-		// if it's readonly mode quit now
-		if app.config.Readonly {
-			return app.NotFound("can't find page")
-		}
-
-		// Allow extensions to handle this page if it's not readonly mode like
-		// opening an editor or something
-		app.Trigger(PageNotFound, page)
-
-		page = DynamicPage{
-			NameVal: page.Name(),
-			RenderFn: func() template.HTML {
-				str := "Page doesn't exist"
-				return template.HTML(str)
-			},
-		}
-	}
-
-	return app.Render("page", Locals{
-		"page": page,
-		"csrf": csrf.Token(r),
-	})
-}
-
 // clearPagesCache clears the pages cache
 func (app *App) clearPagesCache(p Page) {
 	app.pages = nil
-}
-
-// Helper methods that need to be implemented
-func (app *App) ago(t time.Time) string {
-	if app.config.Readonly {
-		return t.Format("Monday 2 January 2006")
-	}
-
-	d := time.Since(t)
-
-	const day = time.Hour * 24
-	const week = day * 7
-	const month = day * 30
-	const year = day * 365
-	const maxPrecision = 2
-
-	var o strings.Builder
-
-	if d.Seconds() < 1 {
-		o.WriteString("Less than a second ")
-	}
-
-	for precision := 0; d.Seconds() > 1 && precision < maxPrecision; precision++ {
-		switch {
-		case d >= year:
-			years := d / year
-			d -= years * year
-			o.WriteString(fmt.Sprintf("%d years ", years))
-		case d >= month:
-			months := d / month
-			d -= months * month
-			o.WriteString(fmt.Sprintf("%d months ", months))
-		case d >= week:
-			weeks := d / week
-			d -= weeks * week
-			o.WriteString(fmt.Sprintf("%d weeks ", weeks))
-		case d >= day:
-			days := d / day
-			d -= days * day
-			o.WriteString(fmt.Sprintf("%d days ", days))
-		case d >= time.Hour:
-			hours := d / time.Hour
-			d -= hours * time.Hour
-			o.WriteString(fmt.Sprintf("%d hours ", hours))
-		case d >= time.Minute:
-			minutes := d / time.Minute
-			d -= minutes * time.Minute
-			o.WriteString(fmt.Sprintf("%d minutes ", minutes))
-		case d >= time.Second:
-			seconds := d / time.Second
-			d -= seconds * time.Second
-			o.WriteString(fmt.Sprintf("%d seconds ", seconds))
-		}
-	}
-
-	o.WriteString("ago")
-
-	return o.String()
-}
-
-// Properties returns a list of properties for a page
-func (app *App) Properties(p Page) map[string]Property {
-
-	ps := map[string]Property{}
-	for _, source := range app.propsSources {
-		for _, pr := range source(p) {
-			ps[pr.Name()] = pr
-		}
-	}
-	return ps
-}
-
-// Links returns a list of links for a Page
-func (app *App) Links(p Page) []Command {
-
-	cmds := []Command{}
-	for _, l := range app.links {
-		cmds = append(cmds, l(p)...)
-	}
-	return cmds
-}
-
-// RenderWidget renders all widget functions registered for a specific widget space
-func (app *App) RenderWidget(s WidgetSpace, p Page) template.HTML {
-
-	w, ok := app.widgets[s]
-	if !ok {
-		return ""
-	}
-
-	var o template.HTML
-	for f := range w.All() {
-		o += f(p)
-	}
-	return o
-}
-
-// Commands returns the list of commands for a page
-func (app *App) Commands(p Page) []Command {
-
-	cmds := []Command{}
-	for _, c := range app.commands {
-		cmds = append(cmds, c(p)...)
-	}
-	return cmds
-}
-
-// QuickCommands returns the list of quick commands for a page
-func (app *App) QuickCommands(p Page) []Command {
-
-	cmds := []Command{}
-	for _, c := range app.quickCommands {
-		cmds = append(cmds, c(p)...)
-	}
-	return cmds
-}
-
-// IsFontAwesome checks if an icon is a FontAwesome icon
-func (app *App) IsFontAwesome(i string) bool {
-	return strings.HasPrefix(i, "fa")
-}
-
-// includeJS adds a JavaScript library URL/path
-func (app *App) includeJS(f string) template.HTML {
-
-	if !slices.Contains(app.js, f) {
-		app.js = append(app.js, f)
-	}
-	return ""
-}
-
-// scripts returns the HTML for all registered JavaScript files
-func (app *App) scripts() template.HTML {
-
-	var b strings.Builder
-	for _, f := range app.js {
-		fmt.Fprintf(&b, `<script src="%s" defer></script>`, f)
-	}
-	return template.HTML(b.String())
-}
-
-// Banner returns the banner image for a page
-func (app *App) Banner(p Page) string {
-	_, a := p.AST()
-	if a == nil {
-		return ""
-	}
-
-	paragraph := a.FirstChild()
-	if paragraph == nil || paragraph.Kind() != ast.KindParagraph {
-		return ""
-	}
-
-	img := paragraph.FirstChild()
-	if img == nil || img.Kind() != ast.KindImage {
-		return ""
-	}
-
-	image, ok := img.(*ast.Image)
-	if !ok {
-		return ""
-	}
-
-	dest := string(image.Destination)
-	if len(dest) == 0 || dest == "#" {
-		return ""
-	}
-
-	if !(path.IsAbs(dest) || strings.HasPrefix(dest, "http")) {
-		d := path.Dir(p.FileName())
-		dest = path.Join("/", d, dest)
-	}
-
-	return dest
-}
-
-// Emoji returns the emoji for a page
-func (app *App) Emoji(p Page) string {
-	_, tree := p.AST()
-	if e, ok := FindInAST[*emojiAst.Emoji](tree); ok && e != nil {
-		return string(e.Value.Unicode)
-	}
-	return ""
-}
-
-// dir returns the directory name
-func (app *App) dir(s string) string {
-	v := path.Dir(s)
-	if v == "." {
-		return ""
-	}
-	return v
-}
-
-// raw returns safe HTML
-func (app *App) raw(i string) template.HTML {
-	return template.HTML(i)
 }
 
 // Redirect returns a redirect response
@@ -885,365 +398,4 @@ func (app *App) NewPage(name string) Page {
 	}
 
 	return &page{name: name}
-}
-
-// Trigger triggers an event
-func (app *App) Trigger(e PageEvent, p Page) {
-	handlers, ok := app.pageEvents[e]
-
-	if !ok {
-		return
-	}
-
-	for _, h := range handlers {
-		if err := h(p); err != nil {
-			slog.Error("Failed to execute handler for event", "event", e, "handler", h, "error", err)
-		}
-	}
-}
-
-// Listen registers an event handler
-func (app *App) Listen(e PageEvent, h PageEventHandler) {
-
-	if _, ok := app.pageEvents[e]; !ok {
-		app.pageEvents[e] = []PageEventHandler{}
-	}
-
-	app.pageEvents[e] = append(app.pageEvents[e], h)
-}
-
-// staticHandler handles static file serving
-func (app *App) staticHandler(r Request) (Output, error) {
-	wd, err := os.Getwd()
-	if err != nil {
-		return nil, err
-	}
-
-	staticFSs := http.FS(
-		priorityFS(
-			append(app.staticDirs, os.DirFS(wd)),
-		),
-	)
-
-	server := http.FileServer(staticFSs)
-
-	cleanPath := path.Clean(r.URL.Path)
-
-	if f, err := staticFSs.Open(cleanPath); err != nil {
-		return nil, err
-	} else {
-		f.Close()
-		return app.Cache(server.ServeHTTP), nil
-	}
-}
-
-// build builds the static site
-func (app *App) build(buildDir string) error {
-	srv := app.server()
-
-	// building Index separately
-	err := app.buildRoute(
-		srv,
-		"/"+app.config.Index,
-		buildDir,
-		path.Join(buildDir, "index.html"),
-	)
-
-	if err != nil {
-		slog.Error("Index Page may not exist, make sure your Index Page exists", "index", app.config.Index, "error", err)
-	}
-
-	errs := app.MapPage(context.Background(), func(p Page) error {
-		err := app.buildRoute(
-			srv,
-			"/"+p.Name(),
-			path.Join(buildDir, p.Name()),
-			path.Join(buildDir, p.Name(), "index.html"),
-		)
-
-		if err != nil {
-			return fmt.Errorf("Failed to process page: %s, error: %w", p.Name(), err)
-		}
-
-		return nil
-	})
-
-	if err := errors.Join(errs...); err != nil {
-		slog.Error(err.Error())
-	}
-
-	// If we render 404 page
-	// Copy 404 page from dest/404/index.html to /dest/404.html
-	if in, err := os.Open(path.Join(buildDir, app.config.NotFoundPage, "index.html")); err == nil {
-		defer in.Close()
-		out, err := os.Create(path.Join(buildDir, "404.html"))
-		if err != nil {
-			slog.Error("Failed to open dest/404.html", "error", err)
-		}
-		defer out.Close()
-		io.Copy(out, in)
-	}
-
-	extensionPageEnclosed := app.extensionPageEnclosed
-	extensionPage := app.extensionPage
-	buildPerms := app.buildPerms
-
-	for route := range extensionPageEnclosed {
-		err := app.buildRoute(
-			srv,
-			route,
-			path.Join(buildDir, route),
-			path.Join(buildDir, route, "index.html"),
-		)
-
-		if err != nil {
-			slog.Error("Failed to process extension page", "route", route, "error", err)
-		}
-	}
-
-	for route := range extensionPage {
-		err := app.buildRoute(
-			srv,
-			route,
-			path.Join(buildDir, path.Dir(route)),
-			path.Join(buildDir, route),
-		)
-
-		if err != nil {
-			slog.Error("Failed to process extension page", "route", route, "error", err)
-		}
-	}
-
-	assets := GetAssets()
-	return fs.WalkDir(assets, ".", func(p string, entry fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-
-		destPath := path.Join(buildDir, p)
-
-		if entry.IsDir() {
-			if err := os.MkdirAll(destPath, buildPerms); err != nil {
-				return err
-			}
-		} else if _, err := os.Stat(destPath); err == nil {
-			slog.Warn("Asset file already exists", "path", destPath)
-		} else {
-			content, err := fs.ReadFile(assets, p)
-			if err != nil {
-				return err
-			}
-
-			if err := os.WriteFile(destPath, content, buildPerms); err != nil {
-				return err
-			}
-		}
-
-		return nil
-	})
-}
-
-// buildRoute builds a single route
-func (app *App) buildRoute(srv *http.Server, route, dir, file string) error {
-	req, err := http.NewRequest(http.MethodGet, route, nil)
-	if err != nil {
-		return err
-	}
-
-	rec := httptest.NewRecorder()
-	srv.Handler.ServeHTTP(rec, req)
-
-	if err := os.MkdirAll(dir, app.buildPerms); err != nil {
-		return err
-	}
-
-	if rec.Result().StatusCode != http.StatusOK {
-		return errors.New(rec.Result().Status)
-	}
-
-	body, err := io.ReadAll(rec.Result().Body)
-	if err != nil {
-		return err
-	}
-	defer rec.Result().Body.Close()
-
-	return os.WriteFile(file, body, app.buildPerms)
-}
-
-// MapPage maps over all pages
-func (app *App) MapPage(ctx context.Context, f func(Page) error) []error {
-	sources := app.sources
-
-	var errs []error
-	for _, source := range sources {
-		source.Each(ctx, func(p Page) {
-			if err := f(p); err != nil {
-				errs = append(errs, err)
-			}
-		})
-	}
-	return errs
-}
-
-// RegisterBuildPage registers a build page
-func (app *App) RegisterBuildPage(p string, encloseInDir bool) {
-
-	if encloseInDir {
-		app.extensionPageEnclosed[p] = true
-	} else {
-		app.extensionPage[p] = true
-	}
-}
-
-// Pages returns all pages
-func (app *App) Pages(ctx context.Context) []Page {
-	if app.pages == nil {
-		app.populatePagesCache(ctx)
-	}
-	pages := app.pages
-	return pages[:]
-}
-
-// EachPage iterates on all available pages
-func (app *App) EachPage(ctx context.Context, f func(Page)) {
-	if app.pages == nil {
-		app.populatePagesCache(ctx)
-	}
-	currentPages := app.pages
-
-	for _, p := range currentPages {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			f(p)
-		}
-	}
-}
-
-// MapPageGeneric maps over all pages with generic return type
-func MapPageGeneric[T any](app *App, ctx context.Context, f func(Page) T) []T {
-	if app.pages == nil {
-		app.populatePagesCache(ctx)
-	}
-	pages := app.pages
-	concurrency := app.concurrency
-
-	grp, ctx := errgroup.WithContext(ctx)
-	grp.SetLimit(concurrency)
-
-	output := make([]T, 0, len(pages))
-	ch := make(chan T, concurrency)
-	done := make(chan bool)
-
-	go func() {
-		for v := range ch {
-			output = append(output, v)
-		}
-		done <- true
-	}()
-
-	for _, p := range pages {
-		select {
-		case <-ctx.Done():
-			break
-		default:
-			grp.Go(func() (err error) {
-				val := f(p)
-				if isNil(val) {
-					return
-				}
-
-				ch <- val
-
-				return
-			})
-		}
-	}
-
-	grp.Wait()
-	close(ch)
-	<-done
-
-	return output
-}
-
-// populatePagesCache populates the pages cache
-func (app *App) populatePagesCache(ctx context.Context) {
-
-	app.pages = make([]Page, 0, 1000)
-	for _, s := range app.sources {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			s.Each(ctx, func(p Page) {
-				app.pages = append(app.pages, p)
-			})
-		}
-	}
-}
-
-// isNil checks if a value is nil
-func isNil[T any](t T) bool {
-	v := reflect.ValueOf(t)
-	kind := v.Kind()
-	// Must be one of these types to be nillable
-	return !v.IsValid() || (kind == reflect.Ptr ||
-		kind == reflect.Interface ||
-		kind == reflect.Slice ||
-		kind == reflect.Map ||
-		kind == reflect.Chan ||
-		kind == reflect.Func) &&
-		v.IsNil()
-}
-
-type priorityItem[T any] struct {
-	Item     T
-	Priority float32
-}
-
-type priorityList[T any] struct {
-	items []priorityItem[T]
-}
-
-func (pl *priorityList[T]) Add(item T, priority float32) {
-	pl.items = append(pl.items, priorityItem[T]{Item: item, Priority: priority})
-	pl.sortByPriority()
-}
-
-func (pl *priorityList[T]) sortByPriority() {
-	sort.Slice(pl.items, func(i, j int) bool {
-		return pl.items[i].Priority < pl.items[j].Priority
-	})
-}
-
-// An iterator over all items
-func (pl *priorityList[T]) All() iter.Seq[T] {
-	return func(yield func(T) bool) {
-		for _, v := range pl.items {
-			if !yield(v.Item) {
-				return
-			}
-		}
-	}
-}
-
-type lastUpdateProp struct{ page Page }
-
-func (a lastUpdateProp) Icon() string { return "fa-solid fa-clock" }
-func (a lastUpdateProp) Name() string { return "modified" }
-func (a lastUpdateProp) Value() any {
-	app := GetApp()
-	return app.ago(a.page.ModTime())
-}
-
-func DefaultProps(p Page) []Property {
-	if p.ModTime().IsZero() {
-		return nil
-	}
-
-	return []Property{
-		lastUpdateProp{p},
-	}
 }
